Adds a new downstream_setup.yml playbook.
This playbook will be used to setup testnodes for downstream
testing by downloading arbitrary yum repos. The repos to install should
be passed in with the --extra-vars option to ansible-playbook.

The yum_repos var can be a string or a list of strings. The value
given will be downloaded into /etc/yum.repos.d

For example:

  --extra-vars '{"yum_repos":["repo1","repo2"]}'
  --extra-vars "yum_repos=repo1"

Signed-off-by: Andrew Schoen <aschoen@redhat.com>